requirements.txt
appdirs==1.4.3
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
cssselect==1.1.0
facebook-scraper==0.1.4
fake-useragent==0.1.11
idna==2.8
igramscraper==0.3.2
lxml==4.4.1
parse==1.12.1
psycopg2-binary==2.8.3
pyee==6.0.0
pyppeteer==0.0.25
pyquery==1.4.0
python-slugify==3.0.2
requests==2.21.0
requests-html==0.10.0
schedule==0.6.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.4
text-unidecode==1.2
tqdm==4.36.1
urllib3==1.24.3
w3lib==1.21.0
websockets==8.0.2
